refactor(workspace): call mfa.WorkspaceErrorCallbacks directly
Moving the callback set into api/mfa left cmd/workspace with a one-line wrapper
around it, while cmd/iam—added in the same branch—calls the exported helper
directly. Two call sites now spell the same thing two ways, so the wrapper is
gone and both update commands call the helper.

Its test moves to api/mfa with it. The behavior it pins is api/mfa's, and from
there it covers cmd/iam's callers too rather than only cmd/workspace's.
