+++ CLAUDE.md
@@ -72,7 +72,7 @@ utils/               # Shared utilities (output, prompts, errors, SSH parsing)
 - **Command registration**: All commands are registered in `cmd/root.go` via `RootCmd.AddCommand()`
 - **API layer**: Each `cmd/` package calls corresponding `api/` package for HTTP requests
 - **SSH-like syntax**: `websh`, `exec`, `cp` support `user@host` syntax via `utils.ParseSSHTarget()`
-- **Error handling**: Common errors (MFA required, username required) are handled via `utils.HandleCommonErrors()` with retry callbacks. `mfa.ErrorCallbacks(ac, retry)` builds the standard callback set for a command that already holds a client—use it instead of writing the whole set out again. Commands whose callbacks genuinely differ construct `utils.ErrorHandlerCallbacks` themselves: `tunnel` creates its client lazily and takes the legacy retry path; `server` needs no username handling; a workspace-level change needs none either and opens a workspace-scoped MFA link, which `mfa.WorkspaceErrorCallbacks(ac, retry)` builds—`cmd/workspace`'s two update commands reach it through a local `errorCallbacks` alias, and `cmd/iam`'s `user role grant`/`revoke` call it directly. Its link is the right one for anything workspace-wide: a role binding names no server, so `mfa.ErrorCallbacks` would hand `GetMFALinkByServerName` an empty name. `utils` cannot host the helper—`api/mfa` and `api/iam` both import `utils`, so `utils` importing either back would be an import cycle
+- **Error handling**: Common errors (MFA required, username required) are handled via `utils.HandleCommonErrors()` with retry callbacks. `mfa.ErrorCallbacks(ac, retry)` builds the standard callback set for a command that already holds a client—use it instead of writing the whole set out again. Commands whose callbacks genuinely differ construct `utils.ErrorHandlerCallbacks` themselves: `tunnel` creates its client lazily and takes the legacy retry path; `server` needs no username handling; a workspace-level change needs none either and opens a workspace-scoped MFA link, which `mfa.WorkspaceErrorCallbacks(ac, retry)` builds and every caller—`cmd/workspace`'s two update commands and `cmd/iam`'s `user role grant`/`revoke`—calls directly. Its link is the right one for anything workspace-wide: a role binding names no server, so `mfa.ErrorCallbacks` would hand `GetMFALinkByServerName` an empty name. `utils` cannot host the helper—`api/mfa` and `api/iam` both import `utils`, so `utils` importing either back would be an import cycle
 - **Workspace identity lives on the client**: `NewAlpaconAPIClient()` pins `BaseURL` and `WorkspaceName` from one `config.LoadConfig()`, and everything below `cmd/`—an MFA link above all—reads `ac.WorkspaceName`. A second read can name a workspace the request is no longer going to: another shell's `alpacon ws use` rewrites the file mid-command, and `alpacon edit` holds a client across a whole editor session. `work-session create --wait --use` blocks on a human approver before `RunUseSession` files the session through `config.SetActiveWorkSessionFor(ac.WorkspaceName, id)`; plain `SetActiveWorkSession` keys off the file's current name and belongs only to a caller with no client, `--unset` (#403). `api/config_boundary_test.go` enforces this over the AST, so a comment naming a reader is not an offender, an aliased import is, and a dot import is rejected on sight. Its reach is `api/` and the reads written there: a helper package that `api/` calls to reach config stays green, and so does a read added under `client/`, the package that holds the pin (#404). What it bars is a named list: `LoadConfig` plus `IsSaaS`, `GetActiveWorkSession`, and `ResolveAuthMethod`, since each one calls it inside. Every other `config` symbol passes—the writes (`CreateConfig`, `DeleteConfig`, `SaveRefreshedAuth0Token`, `SetActiveWorkSession`, `SwitchWorkspace`) and the device-id helpers `api/auth0` calls (`IsValidDeviceID`, `GetOrCreateDeviceID`), which answer nothing about the workspace
 - **Custom flag parsing**: `websh` and `exec` use `DisableFlagParsing: true` and parse flags manually. `exec` supports `--` separator for remote command flags
 - **Shared command execution**: `exec.RunCommandWithRetry()` wraps `event.RunCommand()` + `HandleCommonErrors()` with MFA/retry logic. Used by both `exec` and `websh`
