--- cmd/workspace/error_callbacks.go
@@ -1,11 +1,0 @@
-package workspace
-
-import (
-	"github.com/alpacax/alpacon-cli/api/mfa"
-	"github.com/alpacax/alpacon-cli/client"
-	"github.com/alpacax/alpacon-cli/utils"
-)
-
-func errorCallbacks(ac *client.AlpaconClient, retry func() error) utils.ErrorHandlerCallbacks {
-	return mfa.WorkspaceErrorCallbacks(ac, retry)
-}
