+++ cmd/workspace/workspace_access_control_update.go
@@ -1,6 +1,7 @@
 package workspace
 
 import (
+	"github.com/alpacax/alpacon-cli/api/mfa"
 	"github.com/alpacax/alpacon-cli/api/workspace"
 	"github.com/alpacax/alpacon-cli/client"
 	"github.com/alpacax/alpacon-cli/config"
@@ -41,7 +42,7 @@ Modify the desired fields, save, and close the editor to apply changes.`,
 		var accessControlDetail []byte
 		accessControlDetail, err = workspace.PatchAccessControl(alpaconClient, data)
 		if err != nil {
-			err = utils.HandleCommonErrors(err, "", errorCallbacks(alpaconClient, func() error {
+			err = utils.HandleCommonErrors(err, "", mfa.WorkspaceErrorCallbacks(alpaconClient, func() error {
 				accessControlDetail, err = workspace.PatchAccessControl(alpaconClient, data)
 				return err
 			}))
